% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_vehicle_categories.R
\name{agg_vehicle_categories}
\alias{agg_vehicle_categories}
\title{Mapping of the different vehicle category nomenclature between HERMESv3 and PHEMLight}
\usage{
agg_vehicle_categories(names_to_look_H3, euro_category, names_to_add_PL)
}
\arguments{
\item{names_to_look_H3}{characters to search in the nomenclature to addapt to PHEMLight, can be more than one.
e.g.: "Urban Buses Standard|Coaches Standard|Urban Buses Midi"}

\item{euro_category}{Euro category of the vehicle type to look for. e.g.: Euro II}

\item{names_to_add_PL}{characters to map of PHEMLight nomenclature. e.g: HDV_CB_D_EU2$}
}
\description{
Mapping of the different vehicle category nomenclature between HERMESv3 and PHEMLight
}
