% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_VSP_EF.R
\name{approx_vsp}
\alias{approx_vsp}
\title{Interpolation of calculated VSP with emission factors, per vehicle group}
\usage{
approx_vsp(drive_cycles_vsp, Veh_group)
}
\arguments{
\item{drive_cycles_vsp}{vehicle specific power calculated for all drive cycles}

\item{Veh_group}{vehicle group (PC, BUS or HDV) to calculate its emissions}
}
\value{
instantaneous emissions of NOx, HC, CO, PM, PN and NO for the specified vehicle group [g/h]
}
\description{
Interpolation of calculated VSP with emission factors, per vehicle group
}
