% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_decel_coast.R
\name{calc_Decel_coast_BUS}
\alias{calc_Decel_coast_BUS}
\title{Deceleration coast calculation.
Currently only the deceleration coast for BUS group is needed, since it did not
had an effect on PC and HDV results, and cost computational time.}
\usage{
calc_Decel_coast_BUS(v, grad = 0)
}
\arguments{
\item{v}{instantaneous vehicle speed, gotten from global emissions file}

\item{grad}{road gradient, defaults to 0}

\item{drive_cycles}{drive cycle file of the vehicles simulated}
}
\value{
Deceleration coast value, for each time-step for the vehicle group selected. Depending on the function used
}
\description{
Deceleration coast calculation.
Currently only the deceleration coast for BUS group is needed, since it did not
had an effect on PC and HDV results, and cost computational time.
}
