% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_idle.R
\name{calc_idles}
\alias{calc_idles}
\title{Calculation of idle}
\usage{
calc_idles(idle, fleetshare)
}
\arguments{
\item{idle}{data frame with the idle emission factors for each vehicle group}

\item{fleetshare}{fleetshare of the vehicle fleet at the area of study}
}
\value{
Average idle emission factors for each vehicle group
}
\description{
Calculation of idle
}
