% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_VSP.R
\name{calc_power_PC}
\alias{calc_power_PC}
\title{Calculation of VSP for PC, BUS and HDV}
\usage{
calc_power_PC(v, a, rot_mass, grad = 0)
}
\arguments{
\item{v}{instantaneous vehicle speed}

\item{a}{instantaneous vehicle acceleration}

\item{rot_mass}{instantaneous rotational mass factor}

\item{grad}{road gradient of the time-step}
}
\value{
vehicle power (VSP) of the vehicle group chosen
}
\description{
Calculation of VSP for PC, BUS and HDV
}
