% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rot_mass_factor.R
\name{calc_rot_mass_factor}
\alias{calc_rot_mass_factor}
\title{Calculation of rotational mass factor}
\usage{
calc_rot_mass_factor(veh_class, v)
}
\arguments{
\item{veh_class}{class of vehicle to be determined its rotational mass factor}

\item{v}{instantaneous vehicle speed}
}
\value{
instantaneous rotational mass factor
}
\description{
Calculation of rotational mass factor
}
