% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sum_specific_var.R
\name{calc_sum_specific_var}
\alias{calc_sum_specific_var}
\title{Calculation of the sum of the specific variables of vehicle class}
\usage{
calc_sum_specific_var(vehicles_data, fleetshares)
}
\arguments{
\item{vehicles_data}{vehicles specific parameters}

\item{fleetshare}{fleetshare of the vehicle fleet at the area of study}
}
\value{
average of the vehicle specific parameters
}
\description{
Calculation of the sum of the specific variables of vehicle class
}
