% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wear.R
\name{calc_tyre_wear}
\alias{calc_tyre_wear}
\title{Calculate tyre, break and road wear emissions for all groups of vehicles}
\usage{
calc_tyre_wear(final_t_wear)
}
\arguments{
\item{final_t_wear}{Value of tyre wear that will previously be calculated}

\item{final_b_wear}{Value of break wear that will previously be calculated}

\item{final_s_wear}{Value of surface wear that will previously be calculated}
}
\value{
tyre, road and surface wear emissions
}
\description{
Calculate tyre, break and road wear emissions for all groups of vehicles
}
