% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_veh_lists.R
\name{create_PC_list}
\alias{create_PC_list}
\title{Create lists of PC, BUS and HDV vehicles from the database.}
\usage{
create_PC_list()
}
\arguments{
\item{fleetshare}{fleetshare of the vehicles fleet of the area of study}
}
\value{
list of all vehicles types for each vehicle group
}
\description{
Create lists of PC, BUS and HDV vehicles from the database.
}
