% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rot_mass_factor.R
\name{data_rot_mass_factor}
\alias{data_rot_mass_factor}
\title{Reading of the different values of Rotational Mass factor for the different vehicle classes.}
\usage{
data_rot_mass_factor(veh_files)
}
\arguments{
\item{path_vehicles}{Path were all vehicle data is. Normally within PHEMLight directories}

\item{files_corrected}{list of vehicle files with names already corrected}
}
\value{
rotational mass factor for the different vehicle classes according to the shape of data of each one.
}
\description{
Reading of the different values of Rotational Mass factor for the different vehicle classes.
}
