% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_veh_data.R
\name{def_header}
\alias{def_header}
\title{These are different functions to read vehicle data needed by PHEMLight:
Emissions, vehicle parameters, critical power and idle.}
\usage{
def_header(file_path)
}
\arguments{
\item{file_path}{path of a file to read its header}

\item{path_vehicles}{Path were all vehicle data is. Normally within PHEMLight directories.}

\item{filename}{emissions filename to be read}

\item{crit_Pname}{critical power file to be read}

\item{file_emis}{file with all emission factors}
}
\description{
These are different functions to read vehicle data needed by PHEMLight:
Emissions, vehicle parameters, critical power and idle.
}
