% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_coldstart.R
\name{f_minone}
\alias{f_minone}
\title{Calculate Cold-start emissions}
\usage{
f_minone(thelist)
}
\arguments{
\item{thelist}{list of files you want to test if they are minor than one}

\item{emissions}{general emission file to calculate cold start emissions}

\item{eCold_Hot_P}{Quotient emission factor of Cold/Hot start for Petrol previously calculated}

\item{eCold_Hot_D}{Quotient parameter of Cold/hot start for Diesel previously calculated}
}
\value{
The cold start emissions for all vehicle groups, for Petrol and Diesel
}
\description{
f_minone finds the values =< to 1, of NOx, HC and PM since COPERT V manual states that If Ecold/Ehot < 1;
1 should be used (p.60). For petrol the cold start values depend on the speed of the vehicle,
for diesel it doesn't, hence calc_Cold_Hot_D doesn't requiere imputs
}
