% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_veh_parameters.R
\name{join_veh_variables}
\alias{join_veh_variables}
\title{Join all vehicle parameters in one data frame}
\usage{
join_veh_variables(vehicle_data, fleetshares)
}
\arguments{
\item{vehicle_data}{vehicle parameters previously calculated}

\item{fleetshares}{vehicle fleetshare distribution of the area of study}
}
\value{
all vehicle group parameters joint in a common data frame
}
\description{
Join all vehicle parameters in one data frame
}
