% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mileage_factor.R
\name{mileage_factor_calc}
\alias{mileage_factor_calc}
\title{Function to to apply the mileage correction factor to vehicles
Only path to mileage data is neded.}
\usage{
mileage_factor_calc(path_mileage_data)
}
\arguments{
\item{path_mileage_data}{path of the mileage data file}
}
\description{
Function to to apply the mileage correction factor to vehicles
Only path to mileage data is neded.
}
