% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_to_hermes.R
\name{save_emis_hermes}
\alias{save_emis_hermes}
\title{Save the emissions obtained into R-Line shape.}
\usage{
save_emis_hermes(df, poll)
}
\arguments{
\item{df}{data.frame with emissions, in [g/m/s]}

\item{poll}{pollutant to be saved into R-Line format: NOx, NO, CO, HC, PM}
}
\description{
Save the emissions obtained into R-Line shape.
}
