% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resuspension.R
\name{weight_LDV}
\alias{weight_LDV}
\title{Calculate resuspension. For that it needs to calculate:
Weight of the LDV and PC fleet over the total (weight_LDV and weight_PC respectively)
The emission factors over the entire drive cycles (ef_calc)}
\usage{
weight_LDV(sharefile)
}
\arguments{
\item{sharefile}{file with the fleet share of the city or area of study}

\item{file_emissions}{emission data frame}

\item{ef_constants}{file with the resuspension emission factors}
}
\value{
Resuspension values for all vehicle groups
}
\description{
Calculate resuspension. For that it needs to calculate:
Weight of the LDV and PC fleet over the total (weight_LDV and weight_PC respectively)
The emission factors over the entire drive cycles (ef_calc)
}
